fix: Rename graphql.go to goql.go
