+++ bootstrap.lock
@@ -1,6 +1,6 @@
 # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
 # vim: set syntax=yaml:
 version: v6.10.2
-generated: 2021-05-30T00:05:20Z
+generated: 2021-06-01T16:09:47Z
 versions:
   devbase: v1.3.1
